Add hint parameter to nextToken function to optimize tokenization
The function will first attempt to extract a token of the given type, bypassing the generic matching process.

Most of the gains come from the way number tokens are extracted. Instead of using a regex, the new logic uses `strtod`, which is significantly faster. Since files are mostly numbers this change boosts parsing speed by two orders of magnitude